fixed typo

+++ src/hooks/useDeleteItem.js
@@ -1,19 +1,32 @@
 import { useCallback } from 'react';
+import { useDispatch } from 'react-redux';
 import useRequest from 'wappsto-blanket/hooks/useRequest';
+import useVisible from 'wappsto-blanket/hooks/useVisible';
+import { setItem } from 'wappsto-redux/actions/items';
+import { itemDelete } from '../util/params';
 
 const useDeleteItem = (item) => {
+  const dispatch = useDispatch();
+  const [ confirmVisible, showDeleteConfirmation, hideDeleteConfirmation ] = useVisible(false);
   const { request, send } = useRequest();
+  let type , id;
+  if(item && item.meta){
+    type = item.meta.type;
+    id = item.meta.id;
+  }
 
   const deleteItem = useCallback(() => {
-    if(item && item.meta){
-      send({
+    hideDeleteConfirmation();
+    if(type && id){
+      const requestId = send({
         method: 'DELETE',
-        url: `/${item.meta.type}/${item.meta.id}`
+        url: `/${type}/${id}`
       });
+      dispatch(setItem(itemDelete + id, requestId));
     }
-  }, [item, send]);
+  }, [type, id, send, hideDeleteConfirmation, dispatch]);
 
-  return { deleteItem, request };
+  return { deleteItem, request, confirmVisible, showDeleteConfirmation, hideDeleteConfirmation };
 }
 
 export default useDeleteItem;
